update spring and jetty versions [ci skip]
this is kind of huge, because jetty changed the JSP engine from jasper to apache. this had effects on how we initialize and handle JSPs. additionally, the pre-compilation of JSPs was triggered during the maven compile phase, but the compiled JSP classes were not used during runtime before due to the wrong config and directory structure. therefore, the JSP sources needed to be included so that the compilation kicks in when a page is requested. this is now fixed, we use ahead-of-time compiled JSP classes.