rename db table "user" to "ocpp_tag" and add a "user" table [ci skip]
until now, the user table was for OCPP related data and was good enough. but now that we introduce customer/user related additional information to the system, it would be confusing to keep the old "user" table as it is.

major refactoring is due to reflect the changes. will not compile. do not use.