link various data tables together
for example, on the reservation table, you can click on the ocpp id tag and chargebox id and you will be directed to the details pages of each one. or you can click on the transaction id, and you will be directed to the transactions table.

links from transactions and OCPP/JSON connection status tables are TODO.