handle edge case differently when no conf key is selected
until now, if no configuration key was selected, we sent an empty list 'key' in the request. this is actually perfectly fine (from ocpp spec: "If the list of keys in the request PDU is empty or missing (it is optional)"). but to be extra safe, we omit this element altogether now.