fix ConcurrentTx issue in StartTransactionResponse
problem: we have a db trigger, that sets in_transaction=true of the ocpp tag referenced in a start transaction. since we first insert the transaction, then read the id tag info of the ocpp tag to be set in the response, the status of ocpp tag is set to ConcurrentTx, which is confusing.

this fix only changes the execution order of the both db queries: 1) get id tag info 2) insert transaction