disable ConcurrentTx check
this causes more harm than good. example use case: if the expiry date is set such, that the local cache is invalidated before stop transaction and the charging station sends an authorize in order to stop a transaction, we will send a "ConcurrentTx" for this rfid. Instead of thinking "oh this might be due to the ongoing transaction that the user wants to stop", the charging station might stop further processing. because of this the process breaks and the transaction cannot be stopped.