store reservations with connector id info in database
even though the connector id must be specified for a ReserveNow operation, for whatever reason, we did not store this information in the reservation table. and since charge_box_id and connector_id pair is unique, we can just store connector_pk. with this update, we also present the connector id in the web ui.