improve: store the request in RequestTask instead of handlers
since we can send the _same_ request to multiple charging stations, in this case it would be redundant to store the same object in each handler (one handler per SOAP call i.e. charging station). we can do better by storing it in the RequestTask once and allowing the handlers to retrieve it.