refactor: remove/deprecate jooqConfig as a spring bean
previously, we used to inject jooqConfig into repositories to init DSLContexts at the spot when we run queries. afterwards, we switched to an immutable DSLContext bean that was injected and used. we still left jooqConfig untouched to not to break existing users of it. after more than 6 months after the switch, i guess we can deprecate the old approach.