remember invalid RFID tags of authorize attempts and present in UI
from now on, we store invalid RFID tags in-memory (not persistent!
the tags will vanish if you shut down steve) and present them in
the OCPP tags page. we also specify an upper bound of 1000 tags to be
stored.

the request in #13 brings up storage in database, but we do not
desire to make db schema changes if not absolutely necessary.
keeping them in memory offers a benefit as side effect: forgetting
them after shutdown prevents infinite growing of this list, only
the most recent (and therefore probably relevant) ones are kept.