refactor: start using ocpp version converters
reason: with the increasing number of ocpp versions, it
becomes harder to manage multiple implementations. starting
with this commit, the business logic will only be implemented
for the latest version. data types of earlier versions will
be mapped to the types of the latest version internally and
passed to the impl of the latest version (in this case
1.2 -> 1.5)

todo: this was only for server. implement the same for client.