fix: NPE in issue #38
reason: SendLocalListTask tries to set operation name in the constructor of
the super class, where a call to getRequest method of SendLocalListTask is made.
but, in order to create the request, ocppTagService field has to be set, which
did not happen yet, since we are in the constructor of the super class.
therefore, a NPE occurs.

use a different strategy like parsing the name of the task object. this implies
that we always have to use reasonable task names that reflect the operation name.