insert transaction of unknown of an idTag and block it (closes #13)
reason: Starting with OCPP 1.6, a charging station can be configured to allow
unknown idTags when offline.

Use case: An offline charging station decides to allow an unknown idTag to start
a transaction. Later, when it is online, it sends a StartTransactionRequest with
this idTag. If we do not insert this idTag, the transaction details will not be
inserted into DB and we will lose valuable information.