move logic of db triggers for setting user.inTransaction to java code
reason: with this commit, we reduce the possibility of deadlocks when processing highly
concurrent start- and stopTransaction messages while testing with abusive-charge-point
(https://github.com/chargegrid/abusive-charge-point). the problem was not concurrent access
of the transaction table (as previously incorrectly guessed), but the ocpp tag triggers attached
to transaction table to set ocppTag.inTransaction=true|false after each insert/update. the
deadlocks happen, because
  1) abusive-charge-point sends multiple transaction messages at the same time (millisecond precision),
  2) all these transaction messages use the same id tag,
 => triggers in db try to set ocppTag.inTransaction field of the same ROW at the same TIME.

the solution is to split the previously long db transaction (transaction + ocpp tag table operations)
into two smaller ones, so the locks are released sooner.