Add recommended, additional configuration keys for OCMF
The OCMF specification by SAFE e.V. recommends in the Best Practices part
three new configuration keys for OCPP, see
https://github.com/SAFE-eV/OCMF-Open-Charge-Metering-Format/blob/master/OCMF-de.md

Add these keys to the list, so that users do not need to enter it manually.

It should not harm, since OCPP clients which do not know these keys should
handle it gracefully.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>