Generate endpoint info on About page based on provided HTTP host header
Showing all the interface addresses we're listening on is not helpful
on the About page. However, we serve the About page, we already know
that we are reaching the server using the given hostname in the
corresponding HTTP header, so let's use it to construct the URLs.

This takes also into account that the request/access already passed
proxy server(s), uses DNS CNAMEs etc.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>