Use the default values for input/output buffers
Before it was using a buffer of 8Mb.
This meant that for every message received/sent, a buffer of 8Mb was
being allocated.
This in turn were causing a log of memory pressure and the service would
crash with Out of Memory.
After this change, the IO buffers have a 4Kb size that doesn't create
the same memory pressure.
Even if the buffer is only 4Kb, it still works for messages that are
bigger than 4Kb.