Fix issue related to invalid characters in tags being submitted.
The default behaviour of SteVe is to create a new OCPP tag if the charging point submits an unknown tag. If this tag contains invalid characters, the entry ends up in an invalid state:

- you cannot see it in the OCPP tag overview page (and also isn't displayed in the unknown tag section)
- if you go to the active/past sessions pane, and click through to the tag over there, you cannot edit or delete the tag since it doesn't pass validation.

This bug gets triggered because some charge points send arbitrary text as a tag when FreeCharge mode is enabled. The Webasto Next [1] chargepoint I use triggers this bug, since it uses the #FreeCharging tag ID when freecharging is enabled.

This PR adds the '#' sign to the whitelist of characters allowed, fixing this issue.

[1]: https://charging.webasto.com/int/products/webasto-next/