Propose Docker Compose V2 for bringing up containers
Attempting this with Docker Compose V1 (docker-compose in its latest
release 1.29.2) results in dockerize (the application container entry
point) failing to resolve the DB host. This gives gives the error
message

    Waiting for tcp://mariadb:3306: dial tcp: lookup mariadb on
    127.0.0.11:53: server misbehaving.

in the log of the application container makes it exit after a while.

Bringing up the containers with

    $ docker compose up -d

works like a treat (with docker 25.0.2).