Fix transaction overview page to list all txns instead of just active.
The default view on the Data Management > Transactions page is to only list active transactions. This is annoying, since you most likely want to get an overview of all transactions registered if you go into this view since there is a separate button to only show the active transactions. The code also suggests it was supposed to be the case that all transactions (and not just the active ones) would be shown.

This commit fixes this.