add impl to delete unfinished tasks (#741)
this should be used with care though: if an unfinished task (i.e. the response did not arrive yet) is deleted
prematurely and the response actually would arrive later, we would not know what to do about/with it.

so, use "Delete Unfinished" only if you are sure that these tasks are just zombies.