add userId as filter to ocpp tags, reservations and transactions
the underlying sql logic did not change drastically, since we lookup the ocpp tags that belong to this userId
and use them as filter (like a regular ocpp tag filter). due to this nature, the returned data does not contain
another dimension "userId" and the rendered table does not contain another column to represent it.