refactor charging profile data structures
the whole task is about getting rid of "Map<String, SchedulePeriod>" and using "List<SchedulePeriod>" instead.
map is a weird choice as a data model but made the frontend impl easier when inserting/removing periods. the
other option, a list, has some caveats when working with spring and forms and frontend, but it makes more sense
from a data modeling perspective.