properties/configuration refactor
* translation of properties to native spring boot properties with prefixes server.* and spring.*
* convert SteveConfiguration singleton into spring bean SteveProperties with remainder of steve-specific properties
* during this conversion we lost the property "https.port": spring does not natively support the embedded server
to open a second port/connector. it is still possible programmatically, but i am not sure whether it is worth it, if spring's
default behaviour does not contain in. if this breaks some workflows/setups, we can reconsider in the future.

tests will fail big time. fixes will come in subsequent commit.