fix: mail address handling in GUI after https://github.com/steve-community/steve/pull/1793
the referenced PR refactored the data structure. as a consequence, the path of the recipients field changed.
therefore, the web UI was displaying all emails in one line with commas in-between instead of one email per line.
this is because BatchInsertConverter was applied "recipients" which does not exist anymore.
it should be "mailSettings.recipients".