refactor: merge ChargePointHelperService into ChargePointService
problem during the refactor: spring boot app did not want to start due to cyclic dependencies

   ocppConfiguration
┌─────┐
|  centralSystemService12_SoapServer
↑     ↓
|  centralSystemService16_Service
↑     ↓
|  chargePointService
↑     ↓
|  ocpp12WebSocketEndpoint
└─────┘

chargePointService was injecting ocpp12WebSocketEndpoint (and other versions variants of it) to get the data out of its internal sessionContextStore.
in order to break the cycle, i had to create a standalone SessionContextStoreHolder which creates and holds references to sessionContextStore instances.
there are, then, provided to WebSocketEndpoint instances but also to ChargePointService

(also related to issue https://github.com/steve-community/steve/issues/1831)