bugfix: FROM/TO timestamp filter is not applied ...
... when getting all transactions (regardless of active or stopped)

reason: notice the "return null" in else-branch of the previous code. this was the reason.
with these changes we start applying the timestamp filter only to START_TIMESTAMP column
for all activity types (active / stopped / all). even for stopped transactions. before this, we were
applying FROM/TO to STOP_TIMESTAMP, if we were getting only stopped transactions. i argue
that this is unnecessary nuance. getting rid of this enabled to use an existing generalistic method.