improve the import process performance
reason: current impl was problematic. bulk and batch combination with BATCH_SIZE=1000 meant the following.
in one INSERT statement, we were binding values for 1000 rows. this was not sent to database yet.
then, we were creating another INSERT statement (with also 1000 rows) on client side and another... until we
had 1000 statements to be sent. 1000 statements (batch) with each 1000 rows (bulk) were created before all of this
was sent to database. it created too much memory pressure in steve.

after the changes, each INSERT statement with bind values for 1000 rows will be sent to database. we are also
stopping doing commits after each statement, since all table changes are happening in a ctx.transaction(..) and
there will be a commit at the end anyway.