improve SessionContextStore.getACopy()
we just want an immutable view of the map without copying the underlying data.
copying the map was a defensive decision early in the project since the lookupTable
stores critical info that should not be messed with. current usages of the method
contain collecting statistics about json stations. a read-only view is enough.