remove synchronized blocks and lock Object in AbstractWebSocketEndpoint
the synchronized blocks were there to create an atomic operation of two sub-operations: modify the store by adding/removing
and return the size. SessionContextStore has locks for modifying operations already, such that it makes the lock on caller side
redundant. we adjust modifying operations of SessionContextStore return the size after the operation. with these changes,
AbstractWebSocketEndpoint is simplified by delegating the same semantics to SessionContextStoreImpl.